@@ -1,14 +1,14 @@
 package com.mrcrayfish.goldenhopper.client.screen;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
 import com.mojang.blaze3d.systems.RenderSystem;
+import com.mojang.blaze3d.vertex.PoseStack;
 import com.mrcrayfish.goldenhopper.inventory.container.GoldenHopperContainer;
-import net.minecraft.client.gui.AbstractGui;
-import net.minecraft.client.gui.screen.inventory.ContainerScreen;
-import net.minecraft.entity.player.PlayerInventory;
-import net.minecraft.inventory.container.Slot;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.text.ITextComponent;
+import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
+import net.minecraft.client.renderer.GameRenderer;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.inventory.Slot;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 
@@ -16,44 +16,44 @@ import net.minecraftforge.api.distmarker.OnlyIn;
  * Author: MrCrayfish
  */
 @OnlyIn(Dist.CLIENT)
-public class GoldenHopperScreen extends ContainerScreen<GoldenHopperContainer>
+public class GoldenHopperScreen extends AbstractContainerScreen<GoldenHopperContainer>
 {
     private static final ResourceLocation HOPPER_GUI_TEXTURE = new ResourceLocation("goldenhopper:textures/gui/container/golden_hopper.png");
 
-    public GoldenHopperScreen(GoldenHopperContainer container, PlayerInventory playerInventory, ITextComponent titleIn)
+    public GoldenHopperScreen(GoldenHopperContainer container, Inventory playerInventory, Component titleIn)
     {
         super(container, playerInventory, titleIn);
-        this.ySize = 133;
+        this.imageHeight = 133;
     }
 
     @Override
-    public void render(MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks)
+    public void render(PoseStack poseStack, int mouseX, int mouseY, float partialTicks)
     {
-        this.renderBackground(matrixStack);
-        super.render(matrixStack, mouseX, mouseY, partialTicks);
-        this.func_230459_a_(matrixStack, mouseX, mouseY);
+        this.renderBackground(poseStack);
+        super.render(poseStack, mouseX, mouseY, partialTicks);
+        this.renderTooltip(poseStack, mouseX, mouseY);
     }
 
     @Override
-    protected void drawGuiContainerForegroundLayer(MatrixStack matrixStack, int mouseX, int mouseY)
+    protected void renderLabels(PoseStack poseStack, int mouseX, int mouseY)
     {
-        this.font.drawString(matrixStack, this.title.getString(), 8.0F, 6.0F, 4210752);
-        this.font.drawString(matrixStack, this.playerInventory.getDisplayName().getString(), 8.0F, (float)(this.ySize - 96 + 2), 4210752);
+        this.font.draw(poseStack, this.title.getString(), 8.0F, 6.0F, 4210752);
+        this.font.draw(poseStack, this.playerInventoryTitle, 8.0F, (float)(this.imageHeight - 96 + 2), 4210752);
     }
 
     @Override
-    protected void drawGuiContainerBackgroundLayer(MatrixStack matrixStack, float partialTicks, int mouseX, int mouseY)
+    protected void renderBg(PoseStack poseStack, float partialTicks, int mouseX, int mouseY)
     {
-        RenderSystem.color4f(1.0F, 1.0F, 1.0F, 1.0F);
-        this.minecraft.getTextureManager().bindTexture(HOPPER_GUI_TEXTURE);
-        int startX = (this.width - this.xSize) / 2;
-        int startY = (this.height - this.ySize) / 2;
-        this.blit(matrixStack, startX, startY, 0, 0, this.xSize, this.ySize);
-
-        Slot slot = this.container.getSlot(0);
-        if(!slot.getHasStack())
+        RenderSystem.setShader(GameRenderer::getPositionColorTexShader);
+        RenderSystem.setShaderTexture(0, HOPPER_GUI_TEXTURE);
+        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
+        int startX = (this.width - this.imageWidth) / 2;
+        int startY = (this.height - this.imageHeight) / 2;
+        this.blit(poseStack, startX, startY, 0, 0, this.imageWidth, this.imageHeight);
+        Slot slot = this.menu.getSlot(0);
+        if(!slot.hasItem())
         {
-            blit(matrixStack, this.guiLeft + slot.xPos, this.guiTop + slot.yPos, this.xSize, 0, 16, 16);
+            this.blit(poseStack, this.leftPos + slot.x, this.topPos + slot.y, this.imageWidth, 0, 16, 16);
         }
     }
 }
