@@ -2,12 +2,12 @@ package com.mrcrayfish.goldenhopper.init;
 
 import com.mrcrayfish.goldenhopper.Reference;
 import com.mrcrayfish.goldenhopper.block.GoldenHopperBlock;
-import net.minecraft.block.Block;
-import net.minecraft.block.material.Material;
-import net.minecraft.item.BlockItem;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemGroup;
-import net.minecraftforge.fml.RegistryObject;
+import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.CreativeModeTab;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.material.Material;
+import net.minecraftforge.fmllegacy.RegistryObject;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 
@@ -21,11 +21,11 @@ public class ModBlocks
 {
     public static final DeferredRegister<Block> REGISTER = DeferredRegister.create(ForgeRegistries.BLOCKS, Reference.MOD_ID);
 
-    public static final RegistryObject<Block> GOLDEN_HOPPER = register("golden_hopper", new GoldenHopperBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(2.0F)));
+    public static final RegistryObject<Block> GOLDEN_HOPPER = register("golden_hopper", new GoldenHopperBlock(Block.Properties.of(Material.METAL).strength(2.0F)));
 
     private static <T extends Block> RegistryObject<T> register(String id, T block)
     {
-        return register(id, block, block1 -> new BlockItem(block1, new Item.Properties().group(ItemGroup.REDSTONE)));
+        return register(id, block, block1 -> new BlockItem(block1, new Item.Properties().tab(CreativeModeTab.TAB_REDSTONE)));
     }
 
     private static <T extends Block> RegistryObject<T> register(String id, T block, @Nullable Function<T, BlockItem> supplier)
