@@ -2,9 +2,9 @@ package com.mrcrayfish.goldenhopper.init;
 
 import com.mrcrayfish.goldenhopper.Reference;
 import com.mrcrayfish.goldenhopper.inventory.container.GoldenHopperContainer;
-import net.minecraft.inventory.container.Container;
-import net.minecraft.inventory.container.ContainerType;
-import net.minecraftforge.fml.RegistryObject;
+import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.inventory.MenuType;
+import net.minecraftforge.fmllegacy.RegistryObject;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 
@@ -13,13 +13,13 @@ import net.minecraftforge.registries.ForgeRegistries;
  */
 public class ModContainers
 {
-    public static final DeferredRegister<ContainerType<?>> REGISTER = DeferredRegister.create(ForgeRegistries.CONTAINERS, Reference.MOD_ID);
+    public static final DeferredRegister<MenuType<?>> REGISTER = DeferredRegister.create(ForgeRegistries.CONTAINERS, Reference.MOD_ID);
 
-    public static final RegistryObject<ContainerType<GoldenHopperContainer>> GOLDEN_HOPPER = register("golden_hopper", GoldenHopperContainer::new);
+    public static final RegistryObject<MenuType<GoldenHopperContainer>> GOLDEN_HOPPER = register("golden_hopper", GoldenHopperContainer::new);
 
-    private static <T extends Container> RegistryObject<ContainerType<T>> register(String id, ContainerType.IFactory<T> factory)
+    private static <T extends AbstractContainerMenu> RegistryObject<MenuType<T>> register(String id, MenuType.MenuSupplier<T> factory)
     {
-        return REGISTER.register(id, () -> new ContainerType<>(factory));
+        return REGISTER.register(id, () -> new MenuType<>(factory));
     }
 
 }
