@@ -1,46 +1,45 @@
 package com.mrcrayfish.goldenhopper.item;
 
 import com.mrcrayfish.goldenhopper.entity.GoldenHopperMinecart;
-import net.minecraft.block.AbstractRailBlock;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.DispenserBlock;
-import net.minecraft.dispenser.DefaultDispenseItemBehavior;
-import net.minecraft.dispenser.IBlockSource;
-import net.minecraft.dispenser.IDispenseItemBehavior;
-import net.minecraft.entity.item.minecart.AbstractMinecartEntity;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.ItemUseContext;
-import net.minecraft.item.MinecartItem;
-import net.minecraft.state.properties.RailShape;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.BlockSource;
+import net.minecraft.core.Direction;
+import net.minecraft.core.dispenser.DefaultDispenseItemBehavior;
+import net.minecraft.core.dispenser.DispenseItemBehavior;
 import net.minecraft.tags.BlockTags;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.Direction;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.vehicle.AbstractMinecart;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.context.UseOnContext;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.BaseRailBlock;
+import net.minecraft.world.level.block.DispenserBlock;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.properties.RailShape;
 
 /**
  * Author: MrCrayfish
  */
 public class GoldenHopperMinecartItem extends Item
 {
-    private static final IDispenseItemBehavior MINECART_DISPENSER_BEHAVIOR = new DefaultDispenseItemBehavior()
+    private static final DispenseItemBehavior MINECART_DISPENSER_BEHAVIOR = new DefaultDispenseItemBehavior()
     {
         private final DefaultDispenseItemBehavior behaviourDefaultDispenseItem = new DefaultDispenseItemBehavior();
 
         @Override
-        public ItemStack dispenseStack(IBlockSource source, ItemStack stack)
+        public ItemStack execute(BlockSource source, ItemStack stack)
         {
-            Direction direction = source.getBlockState().get(DispenserBlock.FACING);
-            World world = source.getWorld();
-            double posX = source.getX() + (double) direction.getXOffset() * 1.125D;
-            double posY = Math.floor(source.getY()) + (double) direction.getYOffset();
-            double posZ = source.getZ() + (double) direction.getZOffset() * 1.125D;
-            BlockPos adjacentPos = source.getBlockPos().offset(direction);
-            BlockState adjacentState = world.getBlockState(adjacentPos);
-            RailShape adjacentShape = adjacentState.getBlock() instanceof AbstractRailBlock ? ((AbstractRailBlock) adjacentState.getBlock()).getRailDirection(adjacentState, world, adjacentPos, null) : RailShape.NORTH_SOUTH;
+            Direction direction = source.getBlockState().getValue(DispenserBlock.FACING);
+            Level level = source.getLevel();
+            double posX = source.x() + (double) direction.getStepX() * 1.125D;
+            double posY = Math.floor(source.y()) + (double) direction.getStepY();
+            double posZ = source.z() + (double) direction.getStepZ() * 1.125D;
+            BlockPos adjacentPos = source.getPos().relative(direction);
+            BlockState adjacentState = level.getBlockState(adjacentPos);
+            RailShape adjacentShape = adjacentState.getBlock() instanceof BaseRailBlock ? ((BaseRailBlock) adjacentState.getBlock()).getRailDirection(adjacentState, level, adjacentPos, null) : RailShape.NORTH_SOUTH;
             double yOffset;
-            if(adjacentState.isIn(BlockTags.RAILS))
+            if(adjacentState.is(BlockTags.RAILS))
             {
                 if(adjacentShape.isAscending())
                 {
@@ -53,13 +52,13 @@ public class GoldenHopperMinecartItem extends Item
             }
             else
             {
-                if(!adjacentState.isAir(world, adjacentPos) || !world.getBlockState(adjacentPos.down()).isIn(BlockTags.RAILS))
+                if(!adjacentState.isAir() || !level.getBlockState(adjacentPos.below()).is(BlockTags.RAILS))
                 {
                     return this.behaviourDefaultDispenseItem.dispense(source, stack);
                 }
 
-                BlockState state = world.getBlockState(adjacentPos.down());
-                RailShape shape = state.getBlock() instanceof AbstractRailBlock ? ((AbstractRailBlock) state.getBlock()).getRailDirection(state, world, adjacentPos.down(), null) : RailShape.NORTH_SOUTH;
+                BlockState state = level.getBlockState(adjacentPos.below());
+                RailShape shape = state.getBlock() instanceof BaseRailBlock ? ((BaseRailBlock) state.getBlock()).getRailDirection(state, level, adjacentPos.below(), null) : RailShape.NORTH_SOUTH;
                 if(direction != Direction.DOWN && shape.isAscending())
                 {
                     yOffset = -0.4D;
@@ -70,58 +69,58 @@ public class GoldenHopperMinecartItem extends Item
                 }
             }
 
-            AbstractMinecartEntity minecart = new GoldenHopperMinecart(world, posX, posY + yOffset, posZ);
-            if(stack.hasDisplayName())
+            AbstractMinecart minecart = new GoldenHopperMinecart(level, posX, posY + yOffset, posZ);
+            if(stack.hasCustomHoverName())
             {
-                minecart.setCustomName(stack.getDisplayName());
+                minecart.setCustomName(stack.getHoverName());
             }
 
-            world.addEntity(minecart);
+            level.addFreshEntity(minecart);
             stack.shrink(1);
             return stack;
         }
 
         @Override
-        protected void playDispenseSound(IBlockSource source)
+        protected void playSound(BlockSource source)
         {
-            source.getWorld().playEvent(1000, source.getBlockPos(), 0);
+            source.getLevel().levelEvent(1000, source.getPos(), 0);
         }
     };
 
     public GoldenHopperMinecartItem(Item.Properties builder)
     {
         super(builder);
-        DispenserBlock.registerDispenseBehavior(this, MINECART_DISPENSER_BEHAVIOR);
+        DispenserBlock.registerBehavior(this, MINECART_DISPENSER_BEHAVIOR);
     }
 
     @Override
-    public ActionResultType onItemUse(ItemUseContext context)
+    public InteractionResult useOn(UseOnContext context)
     {
-        World world = context.getWorld();
-        BlockPos pos = context.getPos();
-        BlockState state = world.getBlockState(pos);
-        if(!state.isIn(BlockTags.RAILS))
+        Level level = context.getLevel();
+        BlockPos pos = context.getClickedPos();
+        BlockState state = level.getBlockState(pos);
+        if(!state.is(BlockTags.RAILS))
         {
-            return ActionResultType.FAIL;
+            return InteractionResult.FAIL;
         }
 
-        ItemStack stack = context.getItem();
-        if(!world.isRemote)
+        ItemStack stack = context.getItemInHand();
+        if(!level.isClientSide)
         {
-            RailShape shape = state.getBlock() instanceof AbstractRailBlock ? ((AbstractRailBlock) state.getBlock()).getRailDirection(state, world, pos, null) : RailShape.NORTH_SOUTH;
+            RailShape shape = state.getBlock() instanceof BaseRailBlock ? ((BaseRailBlock) state.getBlock()).getRailDirection(state, level, pos, null) : RailShape.NORTH_SOUTH;
             double yOffset = 0.0D;
             if(shape.isAscending())
             {
                 yOffset = 0.5D;
             }
-            AbstractMinecartEntity minecart = new GoldenHopperMinecart(world, pos.getX() + 0.5, pos.getY() + 0.0625 + yOffset, pos.getZ() + 0.5);
-            if(stack.hasDisplayName())
+            AbstractMinecart minecart = new GoldenHopperMinecart(level, pos.getX() + 0.5, pos.getY() + 0.0625 + yOffset, pos.getZ() + 0.5);
+            if(stack.hasCustomHoverName())
             {
-                minecart.setCustomName(stack.getDisplayName());
+                minecart.setCustomName(stack.getHoverName());
             }
-            world.addEntity(minecart);
+            level.addFreshEntity(minecart);
         }
         stack.shrink(1);
-        return ActionResultType.SUCCESS;
+        return InteractionResult.SUCCESS;
     }
 }
